test: move _writeState transfer coverage into zlib-writestate-transfer.test.js
The regression tests run in isolation alongside the other focused zlib
lifetime tests (zlib-reset-race, zlib-onerror-reentrancy), keeping
zlib.test.js untouched.
